Use ImportError, Swap unittest import order
These are swapped since unittest2 is intended to be used if you're on Python 2. More people will be using Python 3 at this point so it's best to import the Python 3 module and if unable to, it will use the backport.

Most people will be using Python 3 so this will also speed things up somewhat as one less unnecessary (almost expected) try catch will be prevented.